refactor: code organization
