+++ keep-ui/features/workflows/ai-assistant/index.ts
@@ -1,0 +1,1 @@
+export { WorkflowBuilderChatSafe } from "./ui/WorkflowBuilderChatSafe";
