+++ keep-ui/features/workflows/ai-assistant/ui/ai-workflow-placeholder.png
